requirements.txt
weaviate-client>=4.9.5
